Fix hashrate stats displaying.

+++ modules/explorer-core/src/main/scala/org/ergoplatform/explorer/db/models/aggregates/MinerStats.scala
@@ -4,7 +4,7 @@ import org.ergoplatform.explorer.Address
 
 final case class MinerStats(
   minerAddress: Address,
-  totalDifficulties: Long,
+  totalDifficulties: BigDecimal,
   totalTime: Long,
   blocksMined: Int,
   minerName: Option[String]
